Use long for key count in cfstats
patch by marcuse; reviewed by yukim for CASSANDRA-8913

+++ src/java/org/apache/cassandra/tools/NodeTool.java
@@ -832,7 +832,7 @@ public class NodeTool
                     if (offHeapSize != null)
                         System.out.println("\t\tOff heap memory used (total): " + format(offHeapSize, humanReadable));
                     System.out.println("\t\tSSTable Compression Ratio: " + probe.getColumnFamilyMetric(keyspaceName, cfName, "CompressionRatio"));
-                    int numberOfKeys = 0;
+                    long numberOfKeys = 0;
                     for (long keys : (long[]) probe.getColumnFamilyMetric(keyspaceName, cfName, "EstimatedColumnCountHistogram"))
                         numberOfKeys += keys;
                     System.out.println("\t\tNumber of keys (estimate): " + numberOfKeys);
